feat: implement financial dashboard API

+++ apps/server/KpiDashboard/src/Infrastructure/DatabaseSeeder.cs
@@ -47,8 +47,8 @@ public static class DatabaseSeeder
                     var orderDate = new DateTimeOffset(date.Year, date.Month, date.Day, 
                         random.Next(0, 24), random.Next(0, 60), random.Next(0, 60), date.Offset);
 
-                    int valueIncVat = random.Next(100, 5001);
-                    int valueExcVat = (int)(valueIncVat / 1.25);
+                    decimal valueIncVat = random.Next(100, 5001);
+                    decimal valueExcVat = valueIncVat / 1.25m;
 
                     orders.Add(new Order
                     {
